check task will only run small subset of functional tests
Check will only run a small subset of functional tests so that features can be
easier implemented rather than waiting 6 minutes for all functional tests on
each android version to run.

Now only the latest version of android and gradle are used for standard checks
and fullTest will run a variety of tests.